Move hosting tests (#7318)
